feat(mobile): full emoji picker for message reactions
Extract emoji picker (categories, grid, show function) from compose_bar
into shared emoji_picker.dart. Add "+" button to the quick-reaction row
in the message long-press sheet that opens the full picker and fires
addReaction with the selected emoji.
